requirements.txt
Django==2.2.1
django-cors-headers==3.0.1
djangorestframework==3.9.4
djangorestframework-jwt==1.11.0
numpy==1.16.3
pandas==0.24.2
Pillow==6.0.0
psycopg2==2.8.2
PyJWT==1.7.1
python-dateutil==2.8.0
pytz==2019.1
scipy==1.2.1
six==1.12.0
sqlparse==0.3.0
toolz==0.9.0
